Whales_and_Price/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9

numpy==2.1.1
pandas==2.2.3
scikit-learn==1.5.2
imbalanced-learn==0.12.3
xgboost==2.1.1

apscheduler==3.10.4
requests>=2.32.5
python-dotenv==1.0.1
